+++ lib/arcane_helper_utils.dart
@@ -1,6 +1,7 @@
 library arcane_helper_utils;
 
 export "package:arcane_helper_utils/src/extensions/date_time.dart";
+export "package:arcane_helper_utils/src/extensions/dynamic.dart";
 export "package:arcane_helper_utils/src/extensions/list.dart";
 export "package:arcane_helper_utils/src/extensions/string.dart";
 export "package:arcane_helper_utils/src/extensions/string_jwt.dart";